move queries and fetchers to api packages
