+++ packages/api-queries/src/site-deployments.ts
@@ -1,0 +1,14 @@
+import { fetchCodeDeployments, fetchCodeDeploymentRuns } from '@automattic/api-core';
+import { queryOptions } from '@tanstack/react-query';
+
+export const codeDeploymentsQuery = ( siteId: number ) =>
+	queryOptions( {
+		queryKey: [ 'site', siteId, 'code-deployments' ],
+		queryFn: () => fetchCodeDeployments( siteId ),
+	} );
+
+export const codeDeploymentRunsQuery = ( siteId: number, deploymentId: number ) =>
+	queryOptions( {
+		queryKey: [ 'site', siteId, 'code-deployments-runs', deploymentId ],
+		queryFn: () => fetchCodeDeploymentRuns( siteId, deploymentId ),
+	} );
